+++ tests/unit_tests/testcases/app/test_navigationwidget.h
@@ -9,6 +9,7 @@ class TestNavigationWidget : public QObject {
    private Q_SLOTS:
     void testParseDocumentSupportsScriptedHeadings();
     void testParseDocumentIgnoresHeadingsInsideFencedCodeBlocks();
+    void testParseDocumentIgnoresFrontmatterSetextLookalike();
 };
 
 #endif    // TESTNAVIGATIONWIDGET_H