+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Added a button to the **Debug settings** to open the active settings file in an
   external editor on systems where the settings are stored in an editable file
   (for [#3656](https://github.com/pbek/QOwnNotes/issues/3656))
+- Fixed frontmatter entries followed by the closing `---` delimiter being shown
+  as **Setext headings** in the **Navigation panel**
+  (for [#3657](https://github.com/pbek/QOwnNotes/issues/3657))
 
 ## 26.7.3